fix(template-string-parser): allow for escaping of curlies
See steps of `parse a template string` in the spec, which states that
curly brackets should not count as expressions when they're preceeded by
a "reverse solidus" aka backslash.

Note that the spec does not mention what happens to the backslashes,
they are not stripped away or otherwise consumed, so they should appear
in the final token contents.

spec: https://github.com/w3c/webcomponents/blob/gh-pages/proposals/Template-Instantiation.md