fix(NodeTemplatePart): don't get into bad state with replace()
Multiple calls to replace with an empty set of call arguments will cause
the `#parts` array to be empty, which means subsequent calls can throw
with an hard to debug error of `[0].before() is not a function`.

If a call to replace is empty, then it should replace the contents with
an empty text node which it can still use to hook into the DOM to
replace subsequent calls, this prevents the error.